Refactor mock responses
