+++ tests/30_routes/mock_responses/__init__.py
@@ -1,2 +1,2 @@
-from .get_collection import *  # noqa: F401,F403
-from .get_collections import *  # noqa: F401,F403
+from .collections import *  # noqa: F401,F403
+from .collections_collectionId import *  # noqa: F401,F403
