+++ tests/30_routes/mock_responses/collections_collectionId.py
@@ -10,7 +10,7 @@ def get_collection_1_url(base_url) -> str:
 
 
 @pytest.fixture(scope="function")
-def mock_response_get_collection_200(api_collection: ApiCollection, get_collection_1_url: str, httpx_mock: HTTPXMock):
+def mock_response_collections_collectionId_get_200(api_collection: ApiCollection, get_collection_1_url: str, httpx_mock: HTTPXMock):
     httpx_mock.add_response(
         method="GET",
         url=get_collection_1_url,
@@ -19,9 +19,8 @@ def mock_response_get_collection_200(api_collection: ApiCollection, get_collecti
 
 
 @pytest.fixture(scope="function")
-def mock_response_get_collection_404(get_collection_1_url: str, httpx_mock: HTTPXMock):
+def mock_response_collections_collectionId_404(get_collection_1_url: str, httpx_mock: HTTPXMock):
     httpx_mock.add_response(
-        method="GET",
         url=get_collection_1_url,
         status_code=404,
         json={
@@ -37,10 +36,9 @@ def mock_response_get_collection_404(get_collection_1_url: str, httpx_mock: HTTP
 
 
 @pytest.fixture(scope="function")
-def mock_response_get_collection_422(base_url: str, httpx_mock: HTTPXMock):
+def mock_response_collections_collectionId_422(base_url: str, httpx_mock: HTTPXMock):
     get_collection_url = f"{base_url}/collections/f"
     httpx_mock.add_response(
-        method="GET",
         url=get_collection_url,
         status_code=422,
         json={
